Conflicts:
	src/tubes-channel.c
	src/tubes-channel.h

+++ src/tube-dbus.h
@@ -71,7 +71,7 @@ GType gabble_tube_dbus_get_type (void);
 GabbleTubeDBus *gabble_tube_dbus_new (GabbleConnection *conn, TpHandle handle,
     TpHandleType handle_type, TpHandle self_handle, TpHandle initiator,
     const gchar *service, GHashTable *parameters, const gchar *stream_id,
-    guint id, GabbleBytestreamIface *bytestream, GabbleMucChannel *muc,
+    guint64 id, GabbleBytestreamIface *bytestream, GabbleMucChannel *muc,
     gboolean requested);
 
 gboolean gabble_tube_dbus_add_name (GabbleTubeDBus *tube, TpHandle handle,
