+++ src/tube-stream.h
@@ -68,7 +68,7 @@ GType gabble_tube_stream_get_type (void);
 GabbleTubeStream *gabble_tube_stream_new (GabbleConnection *conn,
     TpHandle handle, TpHandleType handle_type, TpHandle self_handle,
     TpHandle initiator, const gchar *service, GHashTable *parameters,
-    guint id, GabbleMucChannel *muc, gboolean requested);
+    guint64 id, GabbleMucChannel *muc, gboolean requested);
 
 gboolean gabble_tube_stream_check_params (TpSocketAddressType address_type,
     const GValue *address, TpSocketAccessControl access_control,
